% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_echarts_theme.R
\name{create_echarts_theme}
\alias{create_echarts_theme}
\title{Création d'un theme echarts4r a partir des preco Akiani}
\usage{
create_echarts_theme(user_theme = "light")
}
\arguments{
\item{user_theme}{light ou dark}
}
\value{
un theme
}
\description{
Création d'un theme echarts4r a partir des preco Akiani
}
\examples{
create_echarts_theme(user_theme = "dark")
create_echarts_theme(user_theme = "light")
}
