% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_svg_path.R
\name{create_svg_path}
\alias{create_svg_path}
\title{Creation un path svg pour les legends echarts4r}
\usage{
create_svg_path(type = "dotted")
}
\arguments{
\item{type}{dotted ou dashed}
}
\value{
un path svg
}
\description{
Creation un path svg pour les legends echarts4r
}
\examples{
create_svg_path(type = "dotted")
create_svg_path(type = "dashed")
}
