% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction_formattage_dates_JS.R
\name{format_horaire}
\alias{format_horaire}
\title{Format horaire a partir d'une date}
\usage{
format_horaire()
}
\value{
un char
}
\description{
Format horaire a partir d'une date
}
\examples{
\dontrun{format_horaire()}
}
