% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction_formattage_dates_JS.R
\name{format_horaire_popup}
\alias{format_horaire_popup}
\title{Format horaire dans un label a partir d'une date}
\usage{
format_horaire_popup()
}
\value{

}
\description{
Format horaire dans un label a partir d'une date
}
\examples{
\dontrun{format_horaire_popup()}
}
