% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction_formattage_dates_JS.R
\name{format_nom_mois_annee_long}
\alias{format_nom_mois_annee_long}
\title{Format ex : Juillet 2020}
\usage{
format_nom_mois_annee_long()
}
\value{

}
\description{
Format ex : Juillet 2020
}
\examples{
\dontrun{format_nom_mois_annee_long()}
}
